design: moved top book repo stuff
