+++ api/src/DataProvider/TopBookCollectionDataProvider.php
@@ -8,7 +8,7 @@ use ApiPlatform\Core\DataProvider\ArrayPaginator;
 use ApiPlatform\Core\DataProvider\ContextAwareCollectionDataProviderInterface;
 use ApiPlatform\Core\DataProvider\RestrictedDataProviderInterface;
 use App\Entity\TopBook;
-use App\Repository\TopBookDataInterface;
+use App\Repository\TopBook\TopBookDataInterface;
 
 final class TopBookCollectionDataProvider implements ContextAwareCollectionDataProviderInterface, RestrictedDataProviderInterface
 {
