+++ api/src/Repository/TopBook/TopBookCachedDataRepository.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Repository;
+namespace App\Repository\TopBook;
 
 use Symfony\Contracts\Cache\CacheInterface;
