+++ api/src/Repository/TopBook/TopBookDataInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Repository;
+namespace App\Repository\TopBook;
 
 /**
  * Interface Component
